lambdas/audit-handler/pyproject.toml
[tool.poetry]
name = "audit_handler"
version = "0.1.0"
description = "AWS Lambda function storing audit events for photos of the application in DynamoDB"
authors = ["Imre Barna <inokone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.34.49"
pydantic = "^2.6.2"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
mypy = "^1.8.0"
bandit = "^1.7.7"
pylint = "^3.1.0"
pytest = "^8.0.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
